Add mixin for recursive delete
