+++ src/lib/y2partitioner/actions/delete_device.rb
@@ -22,6 +22,7 @@
 require "yast"
 require "yast/i18n"
 require "y2partitioner/device_graphs"
+require "y2partitioner/confirm_recursive_delete"
 require "y2storage/filesystems/btrfs"
 require "abstract_method"
 
@@ -35,6 +36,7 @@ module Y2Partitioner
       include Yast::Logger
       include Yast::I18n
       include Yast::UIShortcuts
+      include ConfirmRecursiveDelete
 
       # Constructor
       # @param device [Y2Storage::Device]
@@ -127,74 +129,6 @@ module Y2Partitioner
       # @return [Y2Storage::Md, nil] nil if the device does not belong to a md raid
       def md
         device.md
-      end
-
-      # Devices that depends on the device to delete
-      #
-      # For example, a Vg depends on some partitions used as physical volumes,
-      # so the vg should be deleted when one of that partitions is deleted.
-      #
-      # This method obtains the name of all devices that should be deleted when
-      # the current device is deleted. This info is useful for some confirm messages.
-      #
-      # @see DeleteDisk@confirm
-      #
-      # @return [Array<String>] name of dependent devices
-      def dependent_devices
-        device.descendants.map do |dev|
-          dev.name if dev.respond_to?(:name)
-        end.compact
-      end
-
-      # Helpful method to show a descriptive confirm message with all affected devices
-      #
-      # @see DeleteDisk@confirm
-      def confirm_recursive_delete(devices, headline, label_before, label_after)
-        button_box = ButtonBox(
-          PushButton(Id(:yes), Opt(:okButton), Yast::Label.DeleteButton),
-          PushButton(
-            Id(:no_button),
-            Opt(:default, :cancelButton),
-            Yast::Label.CancelButton
-          )
-        )
-
-        fancy_question(headline,
-          label_before,
-          Yast::HTML.List(devices.sort),
-          label_after,
-          button_box)
-      end
-
-      # @param rich_text [String]
-      # @return [Boolean]
-      def fancy_question(headline, label_before, rich_text, label_after, button_term)
-        display_info = Yast::UI.GetDisplayInfo || {}
-        has_image_support = display_info["HasImageSupport"]
-
-        layout = VBox(
-          VSpacing(0.4),
-          HBox(
-            has_image_support ? Top(Image(Yast::Icon.IconPath("question"))) : Empty(),
-            HSpacing(1),
-            VBox(
-              Left(Heading(headline)),
-              VSpacing(0.2),
-              Left(Label(label_before)),
-              VSpacing(0.2),
-              Left(RichText(rich_text)),
-              VSpacing(0.2),
-              Left(Label(label_after)),
-              button_term
-            )
-          )
-        )
-
-        Yast::UI.OpenDialog(layout)
-        ret = Yast::UI.UserInput
-        Yast::UI.CloseDialog
-
-        ret == :yes
       end
     end
   end
